pyproject.toml
[tool.poetry]
name = "test_task_komtek"
version = "0.1.0"
description = ""
authors = ["ajib6ept <asalbert@narod.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
django-extensions = "^3.2.1"
django-debug-toolbar = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
